test CLI generate from file
