+++ src/lib.rs
@@ -4,6 +4,7 @@
 pub mod cli;
 mod constants;
 pub mod errors;
+pub mod golden;
 pub mod vrf;
 pub mod vrf03;
 pub mod vrf10;
